backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-tika==2.6.0
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4